Resolve app.testing dependency blocking oauth extraction
