+++ app/core/build.gradle.kts
@@ -28,6 +28,7 @@ dependencies {
     implementation(libs.mime4j.dom)
 
     testApi(projects.core.testing)
+    testApi(projects.core.android.testing)
     testImplementation(projects.mail.testing)
     testImplementation(projects.backend.imap)
     testImplementation(projects.mail.protocols.smtp)
