+++ app/core/src/test/java/com/fsck/k9/helper/IdentityHelperTest.kt
@@ -1,10 +1,10 @@
 package com.fsck.k9.helper
 
+import app.k9mail.core.android.testing.RobolectricTest
 import assertk.assertThat
 import assertk.assertions.isEqualTo
 import com.fsck.k9.Account
 import com.fsck.k9.Identity
-import com.fsck.k9.RobolectricTest
 import com.fsck.k9.mail.Address
 import com.fsck.k9.mail.Message
 import com.fsck.k9.mail.Message.RecipientType
