+++ app/core/src/test/java/com/fsck/k9/helper/ListUnsubscribeHelperTest.kt
@@ -1,10 +1,10 @@
 package com.fsck.k9.helper
 
 import androidx.core.net.toUri
+import app.k9mail.core.android.testing.RobolectricTest
 import assertk.assertThat
 import assertk.assertions.isEqualTo
 import assertk.assertions.isNull
-import com.fsck.k9.RobolectricTest
 import com.fsck.k9.mail.internet.MimeMessage
 import org.junit.Test
