+++ app/core/src/test/java/com/fsck/k9/helper/MessageHelperTest.kt
@@ -4,12 +4,12 @@ import android.graphics.Color
 import android.text.SpannableString
 import app.k9mail.core.android.common.contact.Contact
 import app.k9mail.core.android.common.contact.ContactRepository
+import app.k9mail.core.android.testing.RobolectricTest
 import app.k9mail.core.common.mail.EmailAddress
 import app.k9mail.core.common.mail.toEmailAddressOrThrow
 import assertk.assertThat
 import assertk.assertions.isEqualTo
 import assertk.assertions.isInstanceOf
-import com.fsck.k9.RobolectricTest
 import com.fsck.k9.helper.MessageHelper.Companion.toFriendly
 import com.fsck.k9.mail.Address
 import org.junit.Test
