+++ app/core/src/test/java/com/fsck/k9/helper/ReplyToParserTest.java
@@ -1,11 +1,10 @@
 package com.fsck.k9.helper;
 
-
 import java.lang.reflect.Array;
 import java.util.ArrayList;
 
+import app.k9mail.core.android.testing.RobolectricTest;
 import com.fsck.k9.Account;
-import com.fsck.k9.RobolectricTest;
 import com.fsck.k9.helper.ReplyToParser.ReplyToAddresses;
 import com.fsck.k9.mail.Address;
 import com.fsck.k9.mail.Message;
