+++ app/core/src/test/java/com/fsck/k9/mailstore/AttachmentResolverTest.java
@@ -5,7 +5,7 @@ import java.util.Map;
 
 import android.net.Uri;
 
-import com.fsck.k9.RobolectricTest;
+import app.k9mail.core.android.testing.RobolectricTest;
 import com.fsck.k9.mail.BodyPart;
 import com.fsck.k9.mail.Multipart;
 import com.fsck.k9.mail.Part;
