+++ app/core/src/test/java/com/fsck/k9/message/IdentityHeaderBuilderTest.kt
@@ -1,10 +1,10 @@
 package com.fsck.k9.message
 
+import app.k9mail.core.android.testing.RobolectricTest
 import assertk.assertThat
 import assertk.assertions.isGreaterThan
 import com.fsck.k9.Account.QuoteStyle
 import com.fsck.k9.Identity
-import com.fsck.k9.RobolectricTest
 import com.fsck.k9.mail.internet.MimeHeaderChecker
 import com.fsck.k9.mail.internet.TextBody
 import org.junit.Test
