+++ app/core/src/test/java/com/fsck/k9/message/IdentityHeaderParserTest.kt
@@ -1,8 +1,8 @@
 package com.fsck.k9.message
 
+import app.k9mail.core.android.testing.RobolectricTest
 import assertk.assertThat
 import assertk.assertions.contains
-import com.fsck.k9.RobolectricTest
 import com.fsck.k9.helper.toCrLf
 import org.junit.Test
