+++ app/core/src/test/java/com/fsck/k9/message/extractors/AttachmentInfoExtractorTest.java
@@ -1,11 +1,10 @@
 package com.fsck.k9.message.extractors;
 
-
 import android.content.Context;
 import android.net.Uri;
 import androidx.annotation.Nullable;
 
-import com.fsck.k9.RobolectricTest;
+import app.k9mail.core.android.testing.RobolectricTest;
 import com.fsck.k9.mail.Part;
 import com.fsck.k9.mail.internet.MimeBodyPart;
 import com.fsck.k9.mail.internet.MimeHeader;
