+++ core/android/testing/src/main/kotlin/app/k9mail/core/android/testing/RobolectricTest.kt
@@ -1,4 +1,4 @@
-package com.fsck.k9
+package app.k9mail.core.android.testing
 
 import android.app.Application
 import org.junit.runner.RunWith
@@ -6,9 +6,7 @@ import org.robolectric.RobolectricTestRunner
 import org.robolectric.annotation.Config
 
 /**
- * A Robolectric test that does not create an instance of our [Application] class [K9].
- *
- * See also [K9RobolectricTest].
+ * A Robolectric test that does not create an instance of our [Application].
  */
 @RunWith(RobolectricTestRunner::class)
 @Config(application = EmptyApplication::class)
