+++ feature/autodiscovery/providersxml/build.gradle.kts
@@ -9,7 +9,7 @@ dependencies {
 
     implementation(libs.timber)
 
-    testImplementation(projects.app.testing)
+    testImplementation(projects.core.android.testing)
     testImplementation(projects.backend.imap)
     testImplementation(libs.robolectric)
     testImplementation(libs.androidx.test.core)
